Controller Added
I added the controller with the fun jump and a like, demo of the level w/ three platforms. There's lots of easy to access variables for the controller if you click on the player and check the inspector; I made little cool dropdowns and stuff.

If anyone wants to play around with the variables to help find a jump that feels the best, please do so!